Track per-segment usage stats (#6)
When merging individual writes into the primary index, we check if the
index already contains an entry for that key to determine if a previous
value just got displaced by the new write or delete. We'll then
aggregate & transmit these deltas to `SegmentStatsActor`. That actor
keeps a ledger of the total number of "live" and "dead" bytes for each
known segment.

Going forward, this ledger will be the basis to decide if the store must
be compacted, and if yes, which segments should be involved in the
compaction to restore invariants.