Ensure output table has sufficient capacity (#9)
This PR adds a helper class to measure the space requirement for writing
out a given `WriteBatch` instance. Based on that information,
`AllocationActor` tracks how much space is still left in the current
output table, and will roll over to a new output table if the incoming
batch wouldn't fit.