Distribute written data over a series of progressively larger output segments (#4)
This PR provides the initial version of Fugu's partitioning strategy
during writes. The `AllocationActor` will track the total amount of data
already written to the store (currently approximated by the overall
number of change sets seen) and roll over to a fresh output segment
whenever the size limit for the current output segment has been reached.

As this segment size limit increases exponentially with the total amount
of data in the store, the number of segments is bounded by the logarithm
of the total data volume. Thus, assuming carefully chosen parameters for
the relationship, the number of segments (= open OS file handles for
file system persistence) can be kept within reasonable bounds in
practice.

## Open tasks

- [x] Partition based on approximate data volume written to store (=
number of key/value bytes), not number of change sets
- [x] Initialize `AllocationActor` with initial data volume seen during
bootstrapping
- [x] Initialize `WriterActor` with highest generation seen during
bootstrapping