Don't report stats for active output segment (#12)
Up until now, `SegmentStatsTracker` reports stats for all segments
represented in the index, including the current output segment. This is
problematic for balancing because it might trigger a compaction due to a
small volume of data in that segment; worse, including it as a source
segment for compaction would conflict with concurrent writes to the
segment.

Therefore, this PR proposes to amend `SegmentStatsTracker` such that the
current output segment receives special treatment and will not be
reported from `ToImmutable()`, hence will not be passed to
`CompactionActor` when determining whether the store needs to be
rebalanced.