Measure storage utilization as a future trigger for compactions (#13)
Currently, `AllocationActor` tracks the amount of space used in each
segment to decide when to switch to new output segments with
increasingly higher capacity.

This PR extends `CompactionActor` to measure utilization of that
capacity by "live" (vs. "stale") data. When the ratio of "live" data vs.
total capacity drops below a certain threshold, the store can then
identify suitable segments for compaction and copy their "live" contents
to a merged segment (compaction itself to be implemented in a future
PR).