Run index actor during bootstrapping, fixes potential deadlock (#39)
Existing code failed to start `IndexActor` before running bootstrapping.
As a result, when loading data sets with more than 512 individual change
sets, the input queue to `IndexActor` would fill up because no actor was
consuming. This reliably deadlocked the bootstrapper.

To fix this, this PR changes startup order to run `IndexActor` during
bootstrapping so that it will build the index in parallel during
bootstrapping.